Revert "Feat: RoleGameView UI Fix and Refactoring"
